# apps/backend/src/tasks/dto/create-attachment.dto.ts
import { IsInt, IsNotEmpty, IsPositive, IsString, MaxLength } from 'class-validator';

export class CreateAttachmentDto {
  @IsString()
  @IsNotEmpty()
  @MaxLength(255)
  fileName!: string;

  @IsString()
  @IsNotEmpty()
  @MaxLength(255)
  contentType!: string;

  @IsInt()
  @IsPositive()
  fileSize!: number;

  @IsString()
  @IsNotEmpty()
  s3Key!: string;
}