# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/projects/{projectId}/tasks/{taskId}/attachments": {
      "get": {
        "operationId": "TaskAttachmentsController_list",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "taskId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["attachments"],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "operationId": "TaskAttachmentsController_create",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "taskId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAttachmentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": ["attachments"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/projects/{projectId}/tasks/{taskId}/attachments/presign": {
      "post": {
        "operationId": "TaskAttachmentsController_presign",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "taskId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PresignAttachmentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": ["attachments"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    }
  },
  "info": {
    "title": "TaskFlow API",
    "description": "REST API for the TaskFlow collaborative task management platform.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "PresignAttachmentDto": {
        "type": "object",
        "properties": {}
      },
      "CreateAttachmentDto": {
        "type": "object",
        "properties": {}
      }
    }
  }
}